Vendor import of clang trunk r178860: http://llvm.org/svn/llvm-project/cfe/trunk@178860
Former-commit-id: 809500fc2c13c8173a16b052304d983864e4a1e1

+++ test/CodeGen/init.c
@@ -765,7 +765,7 @@ name|x
 block|}
 decl_stmt|;
 comment|// CHECK: @test13
-comment|// CHECK: and i32 {{.*}}, -1024
+comment|// CHECK: and i16 {{.*}}, -1024
 block|}
 end_function
